fix(seo): track owner metadata in sitemap dates
Move the site owner identity into its own source file and include that file in the home and article source paths. This keeps sitemap lastmod values aligned with identity metadata changes without making every page depend on unrelated generator changes.
